모든 클래스를 각 명칭에 맞는 python 파일로 분리
