+++ src/cook.py
@@ -1,20 +1,33 @@
-from dataclasses import dataclass
 
 
 class Cook:
+    def __init__(self):
 
-    def __init__(self, request: dataclass):
-        self.__info = request
-        self.__cooking_time = request.cooking_time
+        self.order = []
+        self.cooks = []
 
-    @property
-    def order_info(self) -> dataclass:
-        return self.__info
+    def waiting_time(self):
+        result = 0
+        if self.order:
+            order, cooks = self.order, self.cooks
+            order.append(0)
+            while order and cooks:
+                cooks.sort()
+                target = cooks.pop(0)
+                result += target
+                cooks = list(map(lambda x: x - target, cooks))
+                cooks.append(order.pop(0))
+        return result
 
-    def get_left_cooking_time(self) -> int:
-        return self.__info.cooking_time
+    def available_new_order(self) -> int:
+        order_num, cook_num = len(self.order), 3 - len(self.cooks)
+        if order_num < cook_num:
+            return order_num
+        return cook_num
 
-    def update(self) -> bool:
-
-        self.__cooking_time -= 1
-        return self.__cooking_time <= 0
+    def update(self):
+        cooks = list(map(lambda x: x - 1, self.cooks))
+        self.cooks = list(filter(lambda x: 0 < x, cooks))
+        pivot = self.available_new_order()
+        self.cooks += self.order[:pivot]
+        self.order = self.order[pivot:]
